+++ R/as_das_df.R
@@ -2,9 +2,9 @@
 #'
 #' Check if an object is of class \code{das_df}, or coerce it if possible.
 #'
-#' @param x n object to be coerced to class \code{das_df}
+#' @param x an object to be coerced to class \code{das_df}
 #'
-#' @details Currently only data frames can be coerced to an object of class \code{das_df}.
+#' @details Only data frames can be coerced to an object of class \code{das_df}.
 #'   If \code{x} does not have column names and classes as specified in \code{\link{das_df-class}},
 #'   then the function returns an error message detailing the first column that does not
 #'   meet the requirements of a \code{das_df} object.
@@ -85,8 +85,8 @@ as_das_df.data.frame <- function(x) {
   # Check that no events are NA
   if (any(is.na(x$Event)))
     stop("The provided data cannot be coerced to an object of class das_df ",
-         "because the following line(s) have NA Event value(s):\n",
-         paste(x$line_num[is.na(x$Event)], collapse = ", "))
+         "because the following have NA Event value(s):\n",
+         .print_file_line(x$file_das, x$line_num, which(is.na(x$Event))))
 
   class(x) <- c("das_df", setdiff(class(x), "das_df"))